Do not kill attaching process when removing task
When a process of service is starting for someone binds it,
before attach application to ActivityManagerService, there
is an activity in the same package is removed from recent app.
The service process will be reported as start timeout because
it has not linkToDeath yet. And also the caller to the service
will not receive onServiceDisconnected

Change-Id: Ic98abc48e8211965a7df97f290836eb1619464df
Signed-off-by: Louis Popi <theh2o64@gmail.com>
Signed-off-by: Louis Popi <theh2o64@gmai.com>
Signed-off-by: Louis Popi <theh2o64@gmail.com>