Built-in mic is not used when USB mic is disconnected
In some cases when the USB microphone is disconnected,
audio stack does not switch to the built-in microphone.
It gets stuck in a state where it still recognizes the
USB mic is still connected. Current device removal
implementation only considers USB output devices such
as headset. The same process should be used for input
USB devices (microphone).

Bug: 24932354
Change-Id: Ic2089ef5a9a318cb47336ade405f79eccd7129f8
Signed-off-by: Alejandro Ochoa <alejandro.ochoa@intel.com>
Signed-off-by: Louis Popi <theh2o64@gmail.com>
Signed-off-by: Louis Popi <theh2o64@gmai.com>
Signed-off-by: Louis Popi <theh2o64@gmail.com>